reorganise upload components
