+++ src/components/Upload/UploadProgress/minimized.tsx
