+++ src/components/Upload/UploadProgress/progressBar.tsx
