+++ src/components/Upload/UploadProgress/styledComponents.tsx
