+++ src/components/Upload/UploadStrategyChoiceModal.tsx
